requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
numpy==1.15.4
opencv-contrib-python==3.4.3.18
opencv-python==3.4.3.18
Pillow==5.3.0
requests==2.20.1
urllib3==1.24.1
